add comments to move ClearJoinHintsWithInvalidPropagationShuttle
