+++ flink-table/flink-table-planner/src/test/java/org/apache/flink/table/planner/hint/ClearJoinHintsWithInvalidPropagationShuttleTestBase.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.alias;
+package org.apache.flink.table.planner.hint;
 
 import org.apache.flink.api.common.RuntimeExecutionMode;
 import org.apache.flink.configuration.Configuration;
